+++ server/src/SocketManager.cpp
@@ -10,7 +10,7 @@ void sock_info_t::init()
 	m_serv_addr.sin_port = htons(6666);
 	m_serv_addr.sin_addr.s_addr = htonl(INADDR_ANY);
 	m_listen_fd = socket(AF_INET, SOCK_STREAM, 0);
-	int16_t opt = 1;
+	int opt = 1;
 	setsockopt(m_listen_fd,SOL_SOCKET, SO_REUSEADDR, &opt, sizeof(opt));
 	if(bind(m_listen_fd, (struct sockaddr*)&m_serv_addr, sizeof(m_serv_addr)) == -1)
     {
@@ -26,9 +26,9 @@ void sock_info_t::init()
 	cout << "server init succeed" << endl;
 }
 
-int16_t sock_info_t::accept_cli()
+int sock_info_t::accept_cli()
 {
-	int16_t conn_fd = accept(m_listen_fd, (struct sockaddr *)&m_cli_addr, &m_len);
+	int conn_fd = accept(m_listen_fd, (struct sockaddr *)&m_cli_addr, &m_len);
 	cout << "port:" << ntohs(get_cli().sin_port) << "; current connect fd = " << conn_fd << "; connected number = " << g_sock_name.size() + 1 << endl;
 	log::log_current_time();
 	g_sock_name[conn_fd] = "";
